# perf: speedup `get_datetime` by ~9.5x