--- a/frappe/tests/test_utils.py
+++ b/frappe/tests/test_utils.py
@@ -667,6 +667,17 @@ class TestDateUtils(IntegrationTestCase):
 		self.assertEqual(get_year_ending(date(2021, 1, 1)), date(2021, 12, 31))
 		self.assertEqual(get_year_ending(date(2021, 1, 31)), date(2021, 12, 31))
 
+	@given(st.datetimes())
+	def test_get_datetime(self, original):
+		parsed = get_datetime(str(original))
+		self.assertEqual(parsed, original)
+
+	@given(st.datetimes(timezones=st.timezones()))
+	def test_get_datetime_tz_aware(self, original):
+		parsed = get_datetime(str(original))
+		original = original.replace(tzinfo=None)
+		self.assertEqual(parsed, original)
+
 	def test_pretty_date(self):
 		from frappe import _