Simple: Use `fromisoformat` written in C instead of Python parser. 

Why this is important?
- Every document has 2 datetime fields, ERPNext has more!
- When you do SQL queries that select datetime fields, this parsing was a significant part of it.


Benchmarks:

```
λ bench  --site bench.localhost run-microbenchmarks --affinity 0 --filter=bench_parse_datetime
```

Before: 11.4 us +- 0.1 us
After: 1.20 us +- 0.01 us